Floats

+++ components/NavigationDrawer/src/private/MDCBottomDrawerHeaderLayer.m
@@ -59,14 +59,14 @@
   [path addLineToPoint:CGPointMake(width - cornerRadius, 0)];
   [path addArcWithCenter:CGPointMake(width - cornerRadius, cornerRadius)
                   radius:cornerRadius
-              startAngle:-M_PI
+              startAngle:-(CGFloat)M_PI
                 endAngle:0
                clockwise:YES];
   [path addLineToPoint:CGPointMake(width, height)];
   [path addLineToPoint:CGPointMake(0, height)];
   [path addArcWithCenter:CGPointMake(cornerRadius, cornerRadius)
                   radius:cornerRadius
-              startAngle:-M_PI
+              startAngle:-(CGFloat)M_PI
                 endAngle:0
                clockwise:YES];
   [path closePath];
